joblyx_server/requirements.txt
fastapi
uvicorn
requests
httpx
python-dotenv
pydantic
pytest
pytest-asyncio
groq
supabase
